Refactor loop CLI commands
