+++ src/heart/cli/commands/run.py
@@ -1,0 +1,31 @@
+from typing import Annotated
+
+import typer
+
+from heart.cli.commands.game_loop import build_game_loop
+from heart.programs.registry import ConfigurationRegistry
+
+
+def run_command(
+    configuration: Annotated[str, typer.Option("--configuration")] = "lib_2025",
+    add_low_power_mode: bool = typer.Option(
+        True, "--add-low-power-mode", help="Add a low power mode"
+    ),
+    x11_forward: bool = typer.Option(
+        False, "--x11-forward", help="Use X11 forwarding for RGB display"
+    ),
+) -> None:
+    registry = ConfigurationRegistry()
+    configuration_fn = registry.get(configuration)
+    if configuration_fn is None:
+        raise Exception(f"Configuration '{configuration}' not found in registry")
+    loop = build_game_loop(x11_forward=x11_forward)
+    configuration_fn(loop)
+
+    ## ============================= ##
+    ## ADD ALL MODES ABOVE THIS LINE ##
+    ## ============================= ##
+    # Retain an empty loop for "lower power" mode
+    if add_low_power_mode:
+        loop.app_controller.add_sleep_mode()
+    loop.start()
